Move directive validation before module parsing
