+++ crates/wgso/src/program.rs
@@ -1,3 +1,4 @@
+use crate::directive::{imports, shader_calls, shader_defs};
 use crate::file::Files;
 use crate::module::Modules;
 use crate::resource::Resources;
@@ -29,6 +30,20 @@ impl Program {
         let mut errors = vec![];
         let directive_rules = crate::directive::load_rules();
         let files = Files::new(root_path, &directive_rules, &mut errors);
+        let directives: Vec<_> = files
+            .iter()
+            .flat_map(|file| file.directives.iter().cloned())
+            .collect();
+        imports::check(&directives, &files, root_path, &mut errors);
+        shader_defs::check(&directives, &mut errors);
+        shader_calls::check(&directives, &mut errors);
+        if !errors.is_empty() {
+            return Self {
+                errors,
+                files,
+                resources: Resources::default(),
+            };
+        }
         let modules = Modules::new(root_path, &files, &mut errors);
         let resources = Resources::new(&files, &modules, &mut errors);
         Self {
