+++ crates/wgso/tests/cases_invalid/wgsl_validation_error/.expected
@@ -1,8 +1,8 @@
-error: expected `shader`, `init`, `run`, `draw` or `import`
- --> ./tests/cases_invalid/wgsl_validation_error/other_error.wgsl:1:2
+error: expected identifier, found ""
+ --> ./tests/cases_invalid/wgsl_validation_error/other_error.wgsl:1:4
   |
-1 | #invalid
-  |  ^
+1 | var
+  |    ^ expected identifier
   |
 error: Function [0] 'increment' is invalid
  --> ./tests/cases_invalid/wgsl_validation_error/root.wgsl:3:1
