+++ crates/wgso/tests/cases_invalid/wgsl_validation_error/other_error.wgsl
@@ -1,1 +1,1 @@
-#invalid
+var
